// src/app/modules/logistica/entrega/retiras/models/pedido.ts
export interface ILogisticaEntregaFormacaoCargaPedido {
  NR_NOTA_FISC: number;
  CD_CHAV_ACES: number;
  TT_VALO_NOTA: number;
  CD_PEDI: string;
  QT_PESO: number;
  DS_LOCA_ENTR: string;
  DS_CIDA: string;
  DS_BAIR: string;
  DS_ESTA: string;
  CD_CLIE: number;
  NM_CLIE: string;
  CD_FILI: number;
  DT_NOTA_FISC: string;
  IN_SELE: boolean;
}